AladinFullstackGateway/src/main/webapp/app/entities/aladin-backend1/heroes/delete/heroes-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { HeroesService } from '../service/heroes.service';

import { HeroesDeleteDialogComponent } from './heroes-delete-dialog.component';

describe('Component Tests', () => {
  describe('Heroes Management Delete Component', () => {
    let comp: HeroesDeleteDialogComponent;
    let fixture: ComponentFixture<HeroesDeleteDialogComponent>;
    let service: HeroesService;
    let mockActiveModal: NgbActiveModal;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
        declarations: [HeroesDeleteDialogComponent],
        providers: [NgbActiveModal],
      })
        .overrideTemplate(HeroesDeleteDialogComponent, '')
        .compileComponents();
      fixture = TestBed.createComponent(HeroesDeleteDialogComponent);
      comp = fixture.componentInstance;
      service = TestBed.inject(HeroesService);
      mockActiveModal = TestBed.inject(NgbActiveModal);
    });

    describe('confirmDelete', () => {
      it('Should call delete service on confirmDelete', inject(
        [],
        fakeAsync(() => {
          // GIVEN
          spyOn(service, 'delete').and.returnValue(of({}));

          // WHEN
          comp.confirmDelete(123);
          tick();

          // THEN
          expect(service.delete).toHaveBeenCalledWith(123);
          expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
        })
      ));

      it('Should not call delete service on clear', () => {
        // GIVEN
        spyOn(service, 'delete');

        // WHEN
        comp.cancel();

        // THEN
        expect(service.delete).not.toHaveBeenCalled();
        expect(mockActiveModal.close).not.toHaveBeenCalled();
        expect(mockActiveModal.dismiss).toHaveBeenCalled();
      });
    });
  });
});
